requirements.txt
asgiref==3.5.0
Django==3.2.13
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
typing-extensions==4.2.0
